Add files via upload
A linked list is a linear data structure in which elements are stored in nodes, and each node points to the next node in the sequence. Linked lists are used to organize and manage a collection of elements. They are particularly useful when dynamic size, efficient insertion and deletion, and constant-time access to elements are required.